Backend/LODD/requirements.txt
absl-py==1.3.0
align==0.0.11
amqp==5.1.1
asgiref==3.5.1
asttokens==2.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
branca==0.5.0
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.24.1.1
colorama==0.4.6
contourpy==1.0.6
cryptography==37.0.2
cycler==0.11.0
Cython==0.29.28
debugpy==1.6.3
decorator==5.1.1
deepface==0.0.75
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.4
django-rest-auth==0.9.5
djangorestframework==3.13.1
dlib==19.24.0
entrypoints==0.4
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet==1.0.5
filelock==3.6.0
fire==0.4.0
Flask==2.2.2
flatbuffers==22.10.26
folium==0.12.1.post1
fonttools==4.33.3
future==0.18.2
gast==0.4.0
gdown==4.5.3
gensim==4.2.0
geographiclib==1.52
geopy==2.2.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
gTTS==2.2.4
h11==0.13.0
h5py==3.7.0
hgtk==0.1.3
idna==3.3
importlib-metadata==5.0.0
imutils==0.5.4
ipykernel==6.17.1
ipython==8.6.0
iso8601==1.1.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter_client==7.4.7
jupyter_core==5.0.0
keras==2.10.0
keras-facenet==0.3.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kombu==5.2.4
lfs==0.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-service==2.4.0
mtcnn==0.1.1
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.1.2.30
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
playsound==1.2.2
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycodestyle==2.8.0
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
redis==4.3.4
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.12.1
serial==0.0.97
sklearn==0.0
smart-open==6.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
stack-data==0.6.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tf-slim==1.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
toml==0.10.2
torch==1.13.0
torchaudio==0.9.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.9
utils==1.0.1
vine==5.0.0
virtualenv==20.14.1
wcwidth==0.2.5
webdriver-manager==3.7.0
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
wsproto==1.1.0
zipp==3.10.0
